requirements.txt
numpy
Pillow
sklearn
scikit-image
django
pandas
